perf(ui): lazy load entity detail components
