+++ openmetadata-ui/src/main/resources/ui/src/components/Entity/EntityLineage/EntityLineage.interface.ts
@@ -13,7 +13,7 @@
 
 import { LoadingState } from 'Models';
 import { ReactNode } from 'react';
-import { Edge as FlowEdge, Node } from 'reactflow';
+import type { Edge as FlowEdge, Node } from 'reactflow';
 import { LineageDirection } from '../../../generated/api/lineage/lineageDirection';
 import { LineageSettings } from '../../../generated/configuration/lineageSettings';
 import { EntityReference } from '../../../generated/entity/type';
