+++ openmetadata-ui/src/main/resources/ui/src/components/Entity/EntityVersionTimeLine/EntityVersionTimeLine.tsx
@@ -11,114 +11,14 @@
  *  limitations under the License.
  */
 import { Button, Col, Divider, Drawer, Row, Tooltip, Typography } from 'antd';
-import classNames from 'classnames';
-import { isEmpty, toString } from 'lodash';
-import { forwardRef, useEffect, useMemo } from 'react';
+import { useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Link } from 'react-router-dom';
 import { useLimitStore } from '../../../context/LimitsProvider/useLimitsStore';
 import { EntityHistory } from '../../../generated/type/entityHistory';
-import { useUserProfile } from '../../../hooks/user-profile/useUserProfile';
-import { formatDateTime } from '../../../utils/date-time/DateTimeUtils';
-import { getEntityName } from '../../../utils/EntityNameUtils';
-import {
-  getSummary,
-  renderVersionButton,
-} from '../../../utils/EntityVersionUtils';
-import { getUserPath } from '../../../utils/RouterUtils';
-import UserPopOverCard from '../../common/PopOverCard/UserPopOverCard';
+import { renderVersionButton } from '../../../utils/EntityVersionUtils';
 import CloseIcon from '../../Modals/CloseIcon.component';
 import './entity-version-timeline.less';
-import {
-  EntityVersionButtonProps,
-  EntityVersionTimelineProps,
-} from './EntityVersionTimeline.interface';
-
-export const VersionButton = forwardRef<
-  HTMLDivElement,
-  EntityVersionButtonProps
->(({ version, onVersionSelect, selected, isMajorVersion, className }, ref) => {
-  const { t } = useTranslation();
-
-  const {
-    updatedBy,
-    version: versionNumber,
-    changeDescription,
-    updatedAt,
-    glossary,
-  } = version;
-  const [, , user] = useUserProfile({
-    permission: true,
-    name: updatedBy,
-  });
-
-  const versionText = `v${parseFloat(versionNumber).toFixed(1)}`;
-
-  return (
-    <div
-      className={classNames(
-        'timeline-content p-b-md cursor-pointer',
-        className
-      )}
-      data-testid={`version-entry-${versionText}`}
-      ref={ref}
-      onClick={() => onVersionSelect(toString(versionNumber))}>
-      <div className="timeline-wrapper">
-        <span
-          className={classNames(
-            'timeline-rounder',
-            {
-              selected,
-            },
-            {
-              major: isMajorVersion,
-            }
-          )}
-          data-testid={`version-selector-${versionText}`}
-        />
-        <span className={classNames('timeline-line')} />
-      </div>
-      <div>
-        <Typography.Text
-          className={classNames('d-flex font-medium', {
-            'text-primary': selected,
-          })}>
-          <span>{versionText}</span>
-          {isMajorVersion ? (
-            <span
-              className="m-l-xs text-xs font-medium text-grey-body tw-bg-tag p-x-xs p-y-xss bg-grey rounded-4"
-              style={{ backgroundColor: '#EEEAF8' }}>
-              {t('label.major')}
-            </span>
-          ) : null}
-        </Typography.Text>
-        <div
-          className={classNames('text-xs font-normal break-all', {
-            'diff-description': selected,
-          })}
-          data-testid="version-change-description">
-          {getSummary({
-            changeDescription: changeDescription,
-            isGlossaryTerm: !isEmpty(glossary),
-          })}
-        </div>
-        <div className="text-xs d-flex gap-1 items-center flex-wrap">
-          <UserPopOverCard
-            className="font-italic"
-            profileWidth={16}
-            userName={updatedBy}>
-            <Link className="thread-author m-r-xss" to={getUserPath(updatedBy)}>
-              {getEntityName(user)}
-            </Link>
-          </UserPopOverCard>
-          <span className="font-medium font-italic version-timestamp">
-            {formatDateTime(updatedAt)}
-          </span>
-        </div>
-      </div>
-    </div>
-  );
-});
+import { EntityVersionTimelineProps } from './EntityVersionTimeline.interface';
 
 const EntityVersionTimeLine: React.FC<EntityVersionTimelineProps> = ({
   versionList = {} as EntityHistory,
