+++ openmetadata-ui/src/main/resources/ui/src/components/Entity/EntityVersionTimeLine/EntityVersionTimeline.interface.ts
@@ -10,6 +10,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
+import type { ReactNode } from 'react';
 import { EntityType } from '../../../enums/entity.enum';
 import { ChangeDescription } from '../../../generated/entity/type';
 import { EntityHistory } from '../../../generated/type/entityHistory';
@@ -34,4 +35,5 @@ export type EntityVersionButtonProps = {
   selected: boolean;
   isMajorVersion: boolean;
   className?: string;
+  summary?: ReactNode;
 };
