+++ openmetadata-ui/src/main/resources/ui/src/components/KnowledgeCenter/KnowledgePageSummary/KnowledgePageSummary.tsx
@@ -12,22 +12,56 @@
  */
 import { Col, Divider, Row, Typography } from 'antd';
 import { isEmpty } from 'lodash';
-import { OwnerLabel } from '../../../components/common/OwnerLabel/OwnerLabel.component';
-import SummaryTagsDescription from '../../../components/common/SummaryTagsDescription/SummaryTagsDescription.component';
-import CommonEntitySummaryInfo from '../../../components/Explore/EntitySummaryPanel/CommonEntitySummaryInfo/CommonEntitySummaryInfo';
+import { lazy, useMemo } from 'react';
+import { Link } from 'react-router-dom';
 import { EntityUnion } from '../../../components/Explore/ExplorePage.interface';
 import {
   KnowledgePage,
   PageType,
   QuickLink,
 } from '../../../interface/knowledge-center.interface';
-
-import { useMemo } from 'react';
-import { Link } from 'react-router-dom';
-import SummaryPanelSkeleton from '../../../components/common/Skeleton/SummaryPanelSkeleton/SummaryPanelSkeleton.component';
 import { DRAWER_NAVIGATION_OPTIONS } from '../../../utils/EntityPureUtils';
 import i18n, { t } from '../../../utils/i18next/LocalUtil';
-import RelatedDataAssets from '../RelatedDataAssets/RelatedDataAssets';
+import withSuspenseFallback from '../../AppRouter/withSuspenseFallback';
+
+const OwnerLabel = withSuspenseFallback(
+  lazy(() =>
+    import('../../../components/common/OwnerLabel/OwnerLabel.component').then(
+      (module) => ({ default: module.OwnerLabel })
+    )
+  )
+);
+
+const SummaryPanelSkeleton = withSuspenseFallback(
+  lazy(
+    () =>
+      import(
+        '../../../components/common/Skeleton/SummaryPanelSkeleton/SummaryPanelSkeleton.component'
+      )
+  )
+);
+
+const SummaryTagsDescription = withSuspenseFallback(
+  lazy(
+    () =>
+      import(
+        '../../../components/common/SummaryTagsDescription/SummaryTagsDescription.component'
+      )
+  )
+);
+
+const RelatedDataAssets = withSuspenseFallback(
+  lazy(() => import('../RelatedDataAssets/RelatedDataAssets'))
+);
+
+const CommonEntitySummaryInfo = withSuspenseFallback(
+  lazy(
+    () =>
+      import(
+        '../../../components/Explore/EntitySummaryPanel/CommonEntitySummaryInfo/CommonEntitySummaryInfo'
+      )
+  )
+);
 
 const KnowledgePageSummary = ({
   entityDetails,
