+++ openmetadata-ui/src/main/resources/ui/src/utils/EntityUtilClassBase.ts
@@ -11,43 +11,23 @@
  *  limitations under the License.
  */
 
-import { ItemType } from 'antd/lib/menu/hooks/useItems';
-import { Operation } from 'fast-json-patch';
+import type { ItemType } from 'antd/lib/menu/hooks/useItems';
+import type { Operation } from 'fast-json-patch';
 import { capitalize } from 'lodash';
-import { FC } from 'react';
-import { NavigateFunction } from 'react-router-dom';
-import DataProductsPage from '../components/DataProducts/DataProductsPage/DataProductsPage.component';
+import type { FC } from 'react';
+import type { NavigateFunction } from 'react-router-dom';
 import { GlobalSettingsMenuCategory } from '../constants/GlobalSettings.constants';
 import {
-  OperationPermission,
   ResourceEntity,
+  type OperationPermission,
 } from '../context/PermissionProvider/PermissionProvider.interface';
 import { EntityType } from '../enums/entity.enum';
 import { SearchIndex } from '../enums/search.enum';
-import { APICollection } from '../generated/entity/data/apiCollection';
-import { Database } from '../generated/entity/data/database';
-import { DatabaseSchema } from '../generated/entity/data/databaseSchema';
-import { ServicesType } from '../interface/service.interface';
-import APICollectionPage from '../pages/APICollectionPage/APICollectionPage';
-import APIEndpointPage from '../pages/APIEndpointPage/APIEndpointPage';
-import ChartDetailsPage from '../pages/ChartDetailsPage/ChartDetailsPage.component';
-import ContainerPage from '../pages/ContainerPage/ContainerPage';
-import DashboardDetailsPage from '../pages/DashboardDetailsPage/DashboardDetailsPage.component';
-import DatabaseDetailsPage from '../pages/DatabaseDetailsPage/DatabaseDetailsPage';
-import DatabaseSchemaPageComponent from '../pages/DatabaseSchemaPage/DatabaseSchemaPage.component';
-import DataModelsPage from '../pages/DataModelPage/DataModelPage.component';
-import DirectoryDetailsPage from '../pages/DirectoryDetailsPage/DirectoryDetailsPage';
-import { VersionData } from '../pages/EntityVersionPage/EntityVersionPage.component';
-import FileDetailsPage from '../pages/FileDetailsPage/FileDetailsPage';
-import MetricDetailsPage from '../pages/MetricsPage/MetricDetailsPage/MetricDetailsPage';
-import MlModelPage from '../pages/MlModelPage/MlModelPage.component';
-import PipelineDetailsPage from '../pages/PipelineDetails/PipelineDetailsPage.component';
-import SearchIndexDetailsPage from '../pages/SearchIndexDetailsPage/SearchIndexDetailsPage';
-import SpreadsheetDetailsPage from '../pages/SpreadsheetDetailsPage/SpreadsheetDetailsPage';
-import StoredProcedurePage from '../pages/StoredProcedure/StoredProcedurePage';
-import TableDetailsPageV1 from '../pages/TableDetailsPageV1/TableDetailsPageV1';
-import TopicDetailsPage from '../pages/TopicDetails/TopicDetailsPage.component';
-import WorksheetDetailsPage from '../pages/WorksheetDetailsPage/WorksheetDetailsPage';
+import type { APICollection } from '../generated/entity/data/apiCollection';
+import type { Database } from '../generated/entity/data/database';
+import type { DatabaseSchema } from '../generated/entity/data/databaseSchema';
+import type { ServicesType } from '../interface/service.interface';
+import type { VersionData } from '../pages/EntityVersionPage/EntityVersionPage.component';
 import { patchApiCollection } from '../rest/apiCollectionsAPI';
 import { patchApiEndPoint } from '../rest/apiEndpointsAPI';
 import { patchApplication } from '../rest/applicationAPI';
@@ -81,6 +61,7 @@ import { ExtraDatabaseDropdownOptions } from './Database/DatabaseDropdownOptions
 import { ExtraDatabaseSchemaDropdownOptions } from './DatabaseSchemaDropdownOptions';
 import { ExtraDatabaseServiceDropdownOptions } from './DatabaseServiceUtils';
 import { getEntityByFqnUtil } from './EntityByFqnUtils';
+import { getEntityDetailComponent as getLazyEntityDetailComponent } from './EntityDetailComponentUtils';
 import { EntityTypeName } from './EntityNameUtils';
 import {
   FormattedAPIServiceType,
@@ -482,51 +463,10 @@ class EntityUtilClassBase {
   }
 
   public getEntityDetailComponent(entityType: string): FC | null {
-    switch (entityType) {
-      case EntityType.DATABASE:
-        return DatabaseDetailsPage;
-      case EntityType.DATABASE_SCHEMA:
-        return DatabaseSchemaPageComponent;
-      case EntityType.PIPELINE:
-        return PipelineDetailsPage;
-      case EntityType.TOPIC:
-        return TopicDetailsPage;
-      case EntityType.DASHBOARD:
-        return DashboardDetailsPage;
-      case EntityType.CHART:
-        return ChartDetailsPage;
-      case EntityType.STORED_PROCEDURE:
-        return StoredProcedurePage;
-      case EntityType.DASHBOARD_DATA_MODEL:
-        return DataModelsPage;
-      case EntityType.MLMODEL:
-        return MlModelPage;
-      case EntityType.CONTAINER:
-        return ContainerPage;
-      case EntityType.SEARCH_INDEX:
-        return SearchIndexDetailsPage;
-      case EntityType.DATA_PRODUCT:
-        return DataProductsPage;
-      case EntityType.TABLE:
-        return TableDetailsPageV1;
-      case EntityType.API_COLLECTION:
-        return APICollectionPage;
-      case EntityType.API_ENDPOINT:
-        return APIEndpointPage;
-      case EntityType.METRIC:
-        return MetricDetailsPage;
-      case EntityType.DIRECTORY:
-        return DirectoryDetailsPage;
-      case EntityType.FILE:
-        return FileDetailsPage;
-      case EntityType.SPREADSHEET:
-        return SpreadsheetDetailsPage;
-      case EntityType.WORKSHEET:
-        return WorksheetDetailsPage;
-
-      default:
-        return null;
-    }
+    // Entity detail pages are large route-level surfaces. Delegate to the lazy
+    // registry so importing EntityUtilClassBase for links/patch APIs does not
+    // pull every detail page into the startup bundle.
+    return getLazyEntityDetailComponent(entityType);
   }
 
   public getResourceEntityFromEntityType(entityType: string): string {
