+++ openmetadata-ui/src/main/resources/ui/src/utils/MetricEntityUtils/MetricUtils.tsx
@@ -12,7 +12,6 @@
  */
 
 import { lazy, Suspense } from 'react';
-import { ActivityFeedTab } from '../../components/ActivityFeed/ActivityFeedTab/ActivityFeedTab.component';
 import { ActivityFeedLayoutType } from '../../components/ActivityFeed/ActivityFeedTab/ActivityFeedTab.interface';
 import withSuspenseFallback from '../../components/AppRouter/withSuspenseFallback';
 import type {
@@ -20,18 +19,49 @@ import type {
   ExtentionEntitiesKeys,
 } from '../../components/common/CustomPropertyTable/CustomPropertyTable.interface';
 import Loader from '../../components/common/Loader/Loader';
-import TabsLabel from '../../components/common/TabsLabel/TabsLabel.component';
-import { GenericTab } from '../../components/Customization/GenericTab/GenericTab';
-import { CommonWidgets } from '../../components/DataAssets/CommonWidgets/CommonWidgets';
-import MetricExpression from '../../components/Metric/MetricExpression/MetricExpression';
-import RelatedMetrics from '../../components/Metric/RelatedMetrics/RelatedMetrics';
-import { SourceType } from '../../components/SearchedData/SearchedData.interface';
+import type { SourceType } from '../../components/SearchedData/SearchedData.interface';
 import { DetailPageWidgetKeys } from '../../enums/CustomizeDetailPage.enum';
 import { EntityTabs, EntityType } from '../../enums/entity.enum';
 import { PageType } from '../../generated/system/ui/page';
-import { WidgetConfig } from '../../pages/CustomizablePage/CustomizablePage.interface';
+import type { WidgetConfig } from '../../pages/CustomizablePage/CustomizablePage.interface';
 import i18n from '../i18next/LocalUtil';
-import { MetricDetailPageTabProps } from './MetricDetailsClassBase';
+import type { MetricDetailPageTabProps } from './MetricDetailsClassBase';
+
+const TabsLabel = withSuspenseFallback(
+  lazy(() => import('../../components/common/TabsLabel/TabsLabel.component'))
+);
+
+const ActivityFeedTab = withSuspenseFallback(
+  lazy(() =>
+    import(
+      '../../components/ActivityFeed/ActivityFeedTab/ActivityFeedTab.component'
+    ).then((module) => ({ default: module.ActivityFeedTab }))
+  )
+);
+
+const GenericTab = withSuspenseFallback(
+  lazy(() =>
+    import('../../components/Customization/GenericTab/GenericTab').then(
+      (module) => ({ default: module.GenericTab })
+    )
+  )
+);
+
+const CommonWidgets = withSuspenseFallback(
+  lazy(() =>
+    import('../../components/DataAssets/CommonWidgets/CommonWidgets').then(
+      (module) => ({ default: module.CommonWidgets })
+    )
+  )
+);
+
+const MetricExpression = withSuspenseFallback(
+  lazy(() => import('../../components/Metric/MetricExpression/MetricExpression'))
+);
+
+const RelatedMetrics = withSuspenseFallback(
+  lazy(() => import('../../components/Metric/RelatedMetrics/RelatedMetrics'))
+);
 
 const CustomPropertyTable = withSuspenseFallback(
   lazy(() =>
