+++ openmetadata-ui/src/main/resources/ui/src/utils/TableColumn.util.tsx
@@ -13,12 +13,9 @@
 import Icon from '@ant-design/icons';
 import { ColumnsType, ColumnType } from 'antd/lib/table';
 import classNames from 'classnames';
+import { lazy } from 'react';
 import { ReactComponent as FilterIcon } from '../assets/svg/ic-filter.svg';
-import { DomainLabel } from '../components/common/DomainLabel/DomainLabel.component';
-import { OwnerLabel } from '../components/common/OwnerLabel/OwnerLabel.component';
-import RichTextEditorPreviewerNew from '../components/common/RichTextEditor/RichTextEditorPreviewNew';
-import DataProductsContainer from '../components/DataProducts/DataProductsContainer/DataProductsContainer.component';
-import TagsViewer from '../components/Tag/TagsViewer/TagsViewer';
+import withSuspenseFallback from '../components/AppRouter/withSuspenseFallback';
 import { TAG_LIST_SIZE } from '../constants/constants';
 import { TABLE_COLUMNS_KEYS } from '../constants/TableKeys.constants';
 import { EntityType } from '../enums/entity.enum';
@@ -32,6 +29,42 @@ import {
   getTagsWithoutTier,
   getTierTags,
 } from './TablePureUtils';
+
+const DomainLabel = withSuspenseFallback(
+  lazy(() =>
+    import('../components/common/DomainLabel/DomainLabel.component').then(
+      (module) => ({ default: module.DomainLabel })
+    )
+  )
+);
+
+const OwnerLabel = withSuspenseFallback(
+  lazy(() =>
+    import('../components/common/OwnerLabel/OwnerLabel.component').then(
+      (module) => ({ default: module.OwnerLabel })
+    )
+  )
+);
+
+const RichTextEditorPreviewerNew = withSuspenseFallback(
+  lazy(
+    () =>
+      import('../components/common/RichTextEditor/RichTextEditorPreviewNew')
+  )
+);
+
+const DataProductsContainer = withSuspenseFallback(
+  lazy(
+    () =>
+      import(
+        '../components/DataProducts/DataProductsContainer/DataProductsContainer.component'
+      )
+  )
+);
+
+const TagsViewer = withSuspenseFallback(
+  lazy(() => import('../components/Tag/TagsViewer/TagsViewer'))
+);
 
 export const columnFilterIcon = (filtered: boolean) => (
   <Icon
